Move derive_enum logic to a separate file
